requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
execnet==1.9.0
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
playwright==1.33.0
pluggy==1.0.0
pyee==9.0.4
pytest==7.3.1
pytest-playwright==0.3.3
pytest-xdist==3.3.1
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.6.0
urllib3==2.0.2
